refactor(edda-serve): split handlers into api/ sub-modules (GH-375)
Extract all HTTP handler functions from the monolithic lib.rs (8.7K lines)
into 12 domain-specific modules under api/:
- auth.rs: device pairing endpoints
- events.rs: health, status, context, decisions, log, note, decide, karvi
- drafts.rs: draft approval workflow
- telemetry.rs: cycle telemetry ingest and stats
- snapshots.rs: snapshot CRUD, village stats, patterns
- analytics.rs: recap, overview, projects
- dashboard.rs: dashboard JSON + compute_attention
- metrics.rs: quality, controls, overview metrics, trends
- policy.rs: scope check, authz, approval, tool-tier
- briefs.rs: task briefs, actors, dashboard HTML
- stream.rs: SSE event stream
- ingestion.rs: ingestion evaluate/records/suggestions

Each module exports routes() for the test router and
public_routes()/protected_routes() for the production server.
lib.rs production code reduced to ~135 lines (entrypoint + mod declarations).
All 106 tests pass, zero clippy warnings.
