+++ crates/edda-serve/src/api/mod.rs
@@ -1,0 +1,12 @@
+pub(crate) mod analytics;
+pub(crate) mod auth;
+pub(crate) mod briefs;
+pub(crate) mod dashboard;
+pub(crate) mod drafts;
+pub(crate) mod events;
+pub(crate) mod ingestion;
+pub(crate) mod metrics;
+pub(crate) mod policy;
+pub(crate) mod snapshots;
+pub(crate) mod stream;
+pub(crate) mod telemetry;
